Smooth grab translate start
it's not finished yet, but this part of the code in MoveSpeedLimiter is what the script was created for. smoothen and limit the speed at which grabbed objects move to and from the player when trying move them around. currently breaks whenever the grab button is used, by multiplying the translate speed by a lot for some reason